More optimization
Changed a lot of datatypes to smaller ones where possible.

+++ UnderworldSketch/Prop.h
@@ -1,18 +1,16 @@
 #ifndef prop_h
 #define prop_h
 
-#include "Direction.h"
 #include "Geo.h"
 
-class Prop
-{
+class Prop {
   private:
     Rect _hitbox;
 
   public:
-    Prop(int x, int y, int width, int height);
-    //Prop(Rect *hitbox);
+    Prop(int x, int y, byte width, byte height);
     Rect* getHitbox();
-    void translate(int x,int y);
+    void translate(int x, int y);
 };
+
 #endif
